% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance.R
\name{attendance}
\alias{attendance}
\alias{legco_attendance}
\title{Attendance of LegCo Members}
\usage{
attendance(
  committee_id = NULL,
  meet_id = NULL,
  member_id = NULL,
  attn = "all",
  from = "1900-01-01",
  to = Sys.Date(),
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_attendance(
  committee_id = NULL,
  meet_id = NULL,
  member_id = NULL,
  attn = "all",
  from = "1900-01-01",
  to = Sys.Date(),
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{committee_id}{The id of a committee, or a vector of ids. If `NULL`,
returns records of meetings from all committee. Defaults to `NULL`.}

\item{meet_id}{The id of a meeting, or a vector of ids. If `NULL`, returns
records of all meetings. Defaults to `NULL`.}

\item{member_id}{The id of a member, or a vector of ids. Defaults to `NULL`.}

\item{attn}{The attendance. `'p'` returns members who were present. `'a'`
returns members who were absent. `'all'` returns all members. Defaults to
`'all'`.}

\item{from}{Only fetch records of meetings on or after this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch records of meetings on or before this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to the current system date.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch attendance record of LegCo committee meetings
}
