% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\alias{legco_bills}
\title{Bills in LegCo}
\usage{
bills(rundown_id = NULL, hansard_id = NULL, section_code = NULL,
  lang = "en", from = "1900-01-01", to = Sys.Date(), floor = FALSE,
  n = 1000, extra_param = NULL, count = FALSE, verbose = TRUE)

legco_bills(rundown_id = NULL, hansard_id = NULL,
  section_code = NULL, lang = "en", from = "1900-01-01",
  to = Sys.Date(), floor = FALSE, n = 1000, extra_param = NULL,
  count = FALSE, verbose = TRUE)
}
\arguments{
\item{rundown_id}{The id of a rundown, or a vector of ids. Defaults to
`NULL`.}

\item{hansard_id}{The id of a hansard file, or a vector of ids. If `NULL`,
returns bills from all hansard files. Defaults to `NULL`.}

\item{section_code}{The section code, or a vector of codes. Defaults to
`NULL`.}

\item{lang}{The language of hansard files to search from. `'en'` returns the
English version. `'zh'` returns the Traditional Chinese version. Defaults
to `'en'`.}

\item{from}{Only fetch results from hansards of meetings on or after this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results from hansards of meetings on or before this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to the current system date.}

\item{floor}{Whether to fetch results from the floor version of the hansard
files. The floor version is the first presented version of hansard file in
the original language delivered by the speakers in LegCo. If `TRUE`, the
language option is ignored. Defaults to `FALSE`.}

\item{n}{The number of records to request. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch bills presented in LegCo council meetings.
}
