% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee.R
\name{committee}
\alias{committee}
\alias{legco_committee}
\title{Committees of LegCo}
\usage{
committee(
  id = NULL,
  code = NULL,
  term_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_committee(
  id = NULL,
  code = NULL,
  term_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{The id of a committee, or a vector of ids. If `NULL`, returns all
committees. Defaults to `NULL`.}

\item{code}{The code of a committee, or a vector of ids. If `NULL`, returns
all committees. Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
committees of all terms.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch basic information of LegCo committees.
}
