% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legco_api.R
\name{legco_api}
\alias{legco_api}
\title{Generic LegCo API Function}
\usage{
legco_api(db, query, n, verbose)
}
\arguments{
\item{db}{The database you wish to access. `'hansard'` for the hansard
database. `'attn'` for the attendance database. `'bill'` for the bills
database. Or the full path name for other database not listed here.}

\item{query}{The query for retrieving data. Should include the data endpoint
and parameters if any.}

\item{n}{The number of entry to fetch. Default to `1000`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
A generic function to access LegCo APIs.
}
