% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legco_section_type.R
\docType{data}
\name{legco_section_type}
\alias{legco_section_type}
\title{Section Types of LegCo Meetings}
\format{A dataframe with 51 rows and 3 variables: \itemize{ \item
  \code{SectionCode}: Unique code(s) for each section type \item
  \code{NameEng}: English name of the section \item \code{NameChi}: Chinese
  name of the section }}
\usage{
legco_section_type
}
\description{
A data frame of possible types of sections in LegCo meetings. For use to look
up the meaning of a section code returned by the API.
}
\examples{
# Look up the meaning of section code "mbm"
legco_section_type[legco_section_type$SectionCode == "mbm", ]

}
\keyword{datasets}
