% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{meeting}
\alias{meeting}
\alias{legco_meeting}
\title{Meetings of LegCo Committees}
\usage{
meeting(
  slot_id = NULL,
  meet_id = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  type = "all",
  term_id = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_meeting(
  slot_id = NULL,
  meet_id = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  type = "all",
  term_id = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{slot_id}{The id of a meeting slot, or a vector of ids. If `NULL`,
returns all meetings. Defaults to `NULL`.}

\item{meet_id}{The id of a meeting, or a vector of ids. If `NULL`, returns
all meetings. Useful for matching meeting with records from the Attendance
Database. Defaults to `NULL`.}

\item{from}{Only fetch results of meetings on or after this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results of meetings on or before this date.  Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to the current system date.}

\item{type}{The type of meeting. If `'open'`, returns open meetings. If
`'closed'`, returns closed meetings. If `'all'`, returns all meetings.
Defaults to `'all'`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
results of all terms. Defaults to `NULL`.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch basic information of LegCo committee meetings.
}
\examples{
\dontrun{
# Fetch all meeting conducted on March 20, 2019
meeting(from = "2019-03-20", to = "2019-03-20")
}

}
