% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_committee.R
\name{meeting_committee}
\alias{meeting_committee}
\alias{legco_meeting_committee}
\title{Committee of Meetings in LegCo}
\usage{
meeting_committee(meet_id = NULL, committee_id = NULL,
  extra_param = NULL, verbose = TRUE)

legco_meeting_committee(meet_id = NULL, committee_id = NULL,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{meet_id}{The id of a LegCo committee meeting, or a vector of ids. If
`NULL`, returns all meetings. Defaults to `NULL`.}

\item{committee_id}{The id of a LegCo committee, or a vector of ids. If
`NULL`, returns committees. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the corresponding committee of meetings in LegCo, or vice versa.
Essentially a function to match Committee ID and Meet ID.
}
