% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_term.R
\name{member_term}
\alias{member_term}
\alias{legco_member_term}
\title{Terms of Members of LegCo}
\usage{
member_term(id = NULL, extra_param = NULL, count = FALSE, verbose = TRUE)

legco_member_term(id = NULL, extra_param = NULL, count = FALSE, verbose = TRUE)
}
\arguments{
\item{id}{The id of a LegCo member, or a vector of ids. If `NULL`, returns
all LegCo members. Defaults to `NULL`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the serving terms in Term ID of LegCo members
}
