% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\alias{legco_membership}
\title{Membership of LegCo Committees}
\usage{
membership(
  member_id = NULL,
  committee_id = NULL,
  term_id = NULL,
  post = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_membership(
  member_id = NULL,
  committee_id = NULL,
  term_id = NULL,
  post = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{member_id}{The id of a LegCo member, or a vector of ids. If `NULL`,
returns results of all LegCo members. Defaults to `NULL`.}

\item{committee_id}{The id of a committee, or a vector of ids. If `NULL`,
returns results of all committees. Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
results of all terms. Defaults to `NULL`.}

\item{post}{The post of a member in the committee, or a vector of posts.
`'President'`, `'Chairman'`, `'Deputy Chairman'`, `'Member'` and
`'Convenor'` return members of the respective post. If `NULL`, returns
members of all posts. Defaults to `NULL`.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the members of LegCo committees.
}
\examples{
\dontrun{
# Fetch members of the Subcommittee on Issues Relating to Bazaars
membership(committee_id = 2704)
}

}
