% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions}
\alias{questions}
\alias{legco_questions}
\title{Questions by LegCo Members}
\usage{
questions(
  rundown_id = NULL,
  speaker_id = NULL,
  type = "all",
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_questions(
  rundown_id = NULL,
  speaker_id = NULL,
  type = "all",
  lang = "en",
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{rundown_id}{The id of a rundown, or a vector of ids. Defaults to
`NULL`.}

\item{speaker_id}{The id of a speaker, or a vector of ids. Defaults to
`NULL`.}

\item{type}{The type of question. `'oral'` returns oral questions.
`'written'` returns written questions. `'all'` returns all questions.
Defaults to `'all'`.}

\item{lang}{The language of hansard files to search from. `'en'` returns the
English version. `'zh'` returns the Traditional Chinese version. Defaults
to `'en'`.}

\item{from}{Only fetch results from hansards of meetings on or after this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results from hansards of meetings on or before this
date.  Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to the current system date.}

\item{floor}{Whether to fetch results from the floor version of the hansard
files. The floor version is the first presented version of hansard file in
the original language delivered by the speakers in LegCo. If `TRUE`, the
language option is ignored. Defaults to `FALSE`.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch questions put to the government by LegCo members in council meetings.
}
