% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundown.R
\name{rundown}
\alias{rundown}
\alias{legco_rundown}
\title{Rundown from LegCo Hansards}
\usage{
rundown(
  rundown_id = NULL,
  hansard_id = NULL,
  speaker_id = NULL,
  lang = "en",
  bookmark_name = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_rundown(
  rundown_id = NULL,
  hansard_id = NULL,
  speaker_id = NULL,
  lang = "en",
  bookmark_name = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  floor = FALSE,
  n = 1000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{rundown_id}{The id of a rundown, or a vector of ids. If `NULL`, returns
results of all rundowns. Defaults to `NULL`.}

\item{hansard_id}{The id of a hansard file, or a vector of ids. If `NULL`,
returns results of all hansard files. Defaults to `NULL`.}

\item{speaker_id}{The id of a speaker at the Legislative Council, or a vector
of ids. If `NULL`, returns all speakers. Defaults to `NULL`.}

\item{lang}{The language of hansard files to search from. `'en'` returns the
English version. `'zh'` returns the Traditional Chinese version. Defaults
to `'en'`.}

\item{bookmark_name}{The bookmark name of the rundown, or a vector of names.
Defaults to `NULL`.}

\item{from}{Only fetch results of meetings on or after this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results of meetings on or before this date.  Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to the current system date.}

\item{floor}{Whether to fetch results from the floor version of the hansard
files. The floor version is the first presented version of hansard file in
the original language delivered by the speakers in LegCo. If `TRUE`, the
language option is ignored. Defaults to `FALSE`.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the rundown from hansard files of LegCo council meetings.
}
\examples{
\dontrun{
# Fetch records of proceedings of the second reading of Smoking (Public Health) (Amendment) Bill 2019
rundown(id = 899628:899649)
}

}
