% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\alias{legco_session}
\title{Session of LegCo}
\usage{
session(
  session_id = NULL,
  term_id = NULL,
  date = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_session(
  session_id = NULL,
  term_id = NULL,
  date = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{session_id}{The id of a session, or a vector of ids. If `NULL`, returns
result of all sessions. Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
results of all terms. Defaults to `NULL`.}

\item{date}{Only fetch the result in which the specified date falls within.
Accepts character values in `'YYYY-MM-DD'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a date with `as.Date()`. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the basic information of LegCo sessions.
}
\examples{
\dontrun{
# Fetch all LegCo sessions of the fifth term
session(term_id = 4)
}

}
