% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakers.R
\name{speakers}
\alias{speakers}
\alias{legco_speakers}
\title{Speakers at LegCo}
\usage{
speakers(
  speaker_id = NULL,
  type = "all",
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_speakers(
  speaker_id = NULL,
  type = "all",
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{speaker_id}{The id of a speaker at the Legislative Council, or a vector
of ids. If `NULL`, returns all speakers. Defaults to `NULL`.}

\item{type}{The position of a speaker. `'all'` returns all speakers. `'PO'`
returns public officers. `'LC'` returns key appointment holders and staff
at LegCo, such as President, Chariman and clerk.`'MB'` returns LegCo
members. Default to `'all'`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the basic information of speakers in LegCo council meetings, including
LegCo members, government officials and Secretariat staff.
}
