% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\alias{legco_term}
\title{Term of LegCo}
\usage{
term(term_id = NULL, date = NULL, extra_param = NULL, verbose = TRUE)

legco_term(term_id = NULL, date = NULL, extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
results of all terms. Defaults to `NULL`.}

\item{date}{Only fetch the result in which the specified date falls within.
Accepts character values in `'YYYY-MM-DD'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a date with `as.Date()`. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the basic information of LegCo terms.
}
\examples{
\dontrun{
# Fetch all LegCo terms
term()
}

}
