% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_record.R
\name{voting_record}
\alias{voting_record}
\alias{legco_voting_record}
\title{Voting Record in LegCo Meetings}
\usage{
voting_record(
  committee = NULL,
  term_id = NULL,
  result = "all",
  vote = "all",
  name_ch = NULL,
  seperate_mechanism = NULL,
  mover_type = "all",
  from = "1900-01-01T00:00:00",
  to = Sys.time(),
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_voting_record(
  committee = NULL,
  term_id = NULL,
  result = "all",
  vote = "all",
  name_ch = NULL,
  seperate_mechanism = NULL,
  mover_type = "all",
  from = "1900-01-01T00:00:00",
  to = Sys.time(),
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{committee}{The name of the committee or subcommittee. Defaults to
`NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
results of all terms. Defaults to `NULL`.}

\item{result}{The voting result. If `'passed'`, returns motions that have
been passed. If `'vetoed'`, returns motions that have been vetoed.If
`'all'`, returns all motions that has been voted in LegCo. Defaults to
`all`.}

\item{vote}{The vote that casted. If `'yes'`, returns only members who casted
affirmative votes. If `'no'`, returns only members who casted negative
votes. If `'abstain'`, returns only members who abstained from voting. If
`'absent'`, returns only members who were absent. If `'present'`, returns
only members who were present and did not vote (e.g. President). If
`'all'`, returns all votes. Defaults to `'all'`.}

\item{name_ch}{The name of a LegCo member, or a vector of names. If `NULL`,
returns voting records of all members. Defaults to `NULL`.}

\item{seperate_mechanism}{Only fetch votes that were counted with the vote
seperate mechanism, i.e. requiring majority in both geographical and
functional constituencies to pass. If `NULL`, returns all votes regardless
of the vote counting mechanism used. Defaults to `NULL`.}

\item{mover_type}{The type of motion being put on vote. If `'PO'`, returns
votes on government motions only. If `'MB'`, returns votes on members'
motions only. If `'all'`, returns votes on all motions. Defaults to
`'all'`.}

\item{from}{Only fetch results of meetings on or after this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results of meetings on or before this date.  Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to the current system date.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch voting records of LegCo council, the Finance Committee and its
subcommittees and the House Committee meetings.
}
\examples{
\dontrun{
# Fetch how members voted the motion on "Abolishing the MPF Offsetting Mechanism" on November 11, 2016
voting_record(committee = "Council Meeting",
              from = "2016-11-16T13:51:53",
              to = "2016-11-16T13:51:53")
}

}
