% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_records.R
\name{voting_records}
\alias{voting_records}
\alias{legco_voting_records}
\title{Voting Records in LegCo Meetings}
\usage{
voting_records(committee = NULL, result = "all", name_ch = NULL,
  from = "1900-01-01", to = Sys.Date(), n = 10000,
  extra_param = NULL, verbose = TRUE)

legco_voting_records(committee = NULL, result = "all",
  name_ch = NULL, from = "1900-01-01", to = Sys.Date(), n = 10000,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{committee}{The name of the committee or subcommittee. Defaults to
`NULL`.}

\item{result}{The voting result. `'passed'` returns motions that have been
passed. `'vetoed'` returns motions that have been vetoed. `'all'` returns
all motions that has been voted in LegCo. Defaults to `all`.}

\item{name_ch}{The name of a LegCo member, or a vector of names. If `NULL`,
returns voting records  of all members. Defaults to `NULL`.}

\item{from}{Only fetch results of meetings on or after this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results of meetings on or before this date. Accepts
character values in `'YYYY-MM-DD'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
date with `as.Date()`. Defaults to the current system date.}

\item{n}{The number of records to request. Defaults to `10000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch voting records from LegCo Council, the Finance Committee and its
subcommittees and the House Committee meetings.
}
