% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{addresses}
\alias{addresses}
\alias{legco_addresses}
\title{Addresses Delivered in LegCo}
\usage{
addresses(hansard_id = NULL, rundown_id = NULL, lang = "en",
  from = "1900-01-01", to = Sys.Date(), floor = FALSE, n = 1000,
  extra_param = NULL, verbose = TRUE)

legco_addresses(hansard_id = NULL, rundown_id = NULL, lang = "en",
  from = "1900-01-01", to = Sys.Date(), floor = FALSE, n = 1000,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{hansard_id}{The id of a hansard file, or a vector of ids. If `NULL`,
returns addresses from all hansard files. Defaults to `NULL`.}

\item{rundown_id}{The id of a rundown, or a vector of ids. Defaults to
`NULL`.}

\item{lang}{The language of hansard files to search from. `'en'` returns the
English version. `'zh'` returns the Traditional Chinese version. Defaults
to `'en'`.}

\item{from}{Only fetch results from hansards of meetings on or after this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only fetch results from hansards of meetings on or before this
date. Accepts character values in `'YYYY-MM-DD'` format, and objects of
class `Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be
coerced to a date with `as.Date()`. Defaults to the current system date.}

\item{floor}{Whether to fetch results from the floor version of the hansard
files. The floor version is the first presented version of hansard file in
the original language delivered by the speakers in LegCo. If `'TRUE'`, the
language option is ignored. Defaults to `FALSE`.}

\item{n}{The number of records to request. Defaults to `1000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch addresses by LegCo members and government officials when tabling papers
and reports in LegCo council meetings.
}
