% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee.R
\name{committee}
\alias{committee}
\alias{legco_committee}
\title{Committees of LegCo}
\usage{
committee(id = NULL, code = NULL, term_id = NULL,
  extra_param = NULL, verbose = TRUE)

legco_committee(id = NULL, code = NULL, term_id = NULL,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{id}{The id of a committee, or a vector of ids. If `NULL`, returns all
committees. Defaults to `NULL`.}

\item{code}{The code of a committee, or a vector of ids. If `NULL`, returns
all committees. Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
committees of all terms.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}

\item{lang}{The language. `'en'` returns result in English. `'zh'` returns
result in Traditional Chinese. Defaults to `'en'`.}
}
\description{
Fetch basic information of LegCo committees.
}
