% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legco-package.R
\docType{package}
\name{legco}
\alias{legco}
\alias{legco-package}
\title{legco: Provides Easy Downloading Capabilities for the UK Parliament API}
\description{
Provides functions to request data from the Hong Kong Legislative Council APIs.
}
\details{
For more details please see the API documentation on <https://www.legco.gov.hk/odata/english/hansard-db.html>.

This package is in no way officially related to or endorsed by the Legislative Council of Hong Kong.
}
