% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legco_api.R
\name{legco_api}
\alias{legco_api}
\title{Generic LegCo API Function}
\usage{
legco_api(db, query, n = 1000, verbose = TRUE)
}
\arguments{
\item{db}{The database you wish to access. `'hansard'` for the hansard
database. `'attn'` for the attendance database. `'bill'` for the bills
database. `'schedule'` for the schedule database. Or the path name for
databases not listed here (i.e. the string between the domain name and data
endpoints, e.g. `'OpenData/HansardDB'`).}

\item{query}{The query for retrieving data. Should include the data endpoint
and parameters if any.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
A generic function to access LegCo APIs.
}
