% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legco_section_type.R
\docType{data}
\name{legco_section_type}
\alias{legco_section_type}
\title{Section Types of LegCo Meetings}
\format{A dataframe with 35 rows and 3 variables:
\describe{
  \item{SectionCode}{Unique code(s) for each section type}
  \item{NameChi}{Chinese name of the section}
  \item{NameEng}{English name of the section}
}}
\usage{
legco_section_type
}
\description{
A data frame of possible types of sections in LegCo meetings.
For use to look up the meaning of a section code returned by the API.
}
\keyword{datasets}
