% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{meeting}
\alias{meeting}
\alias{legco_meeting}
\title{Meetings of LegCo Committees}
\usage{
meeting(
  id = NULL,
  meet_id = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  type = "all",
  term_id = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_meeting(
  id = NULL,
  meet_id = NULL,
  from = "1900-01-01",
  to = Sys.Date(),
  type = "all",
  term_id = NULL,
  n = 10000,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{id}{The id of a meeting slot, or a vector of ids. If `NULL`, returns
all meetings. Defaults to `NULL`.}

\item{meet_id}{The id of a meeting, or a vector of ids. If `NULL`, returns
all meetings. Useful for matching meeting with records from the Attendance
Database. Defaults to `NULL`.}

\item{from}{Only fetch meetings on or after this date. Accepts character
values in `'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`,
`POSIXct`, `POSIXlt` or anything else that can be coerced to a date with
`as.Date()`. Defaults to `'1900-01-01'`.}

\item{to}{Only meetings on or before this date. Accepts character values in
`'YYYY-MM-DD'` format, and objects of class `Date`, `POSIXt`, `POSIXct`,
`POSIXlt` or anything else that can be coerced to a date with `as.Date()`.
Defaults to the current system date.}

\item{type}{The type of meeting. If `all`, returns membership of committees
of all terms.}

\item{term_id}{The id of a term, or a vector of ids. If `'open'`, returns
open meetings. If `'closed'`, returns closed meetings. If `'all'`, returns
all meetings. Defaults to `'all'`.}

\item{n}{The number of record to fetch. Defaults to `1000`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch basic information of LegCo committee meetings.
}
