% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_committee.R
\name{meeting_committee}
\alias{meeting_committee}
\alias{legco_meeting_committee}
\title{Committee of Meetings in LegCo}
\usage{
meeting_committee(
  slot_id = NULL,
  meet_id = NULL,
  committee_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_meeting_committee(
  slot_id = NULL,
  meet_id = NULL,
  committee_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{slot_id}{The id of a LegCo meeting slot, or a vector of ids. If `NULL`,
returns all meetings. Defaults to `NULL`.}

\item{meet_id}{The id of a meeting, or a vector of ids. If `NULL`, returns
all meetings. Useful for matching meeting with records from the Attendance
Database. Defaults to `NULL`.}

\item{committee_id}{The id of a LegCo committee, or a vector of ids. If
`NULL`, returns committees. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch all the meeting slots of a given committee(s), or the committee of a
given meeting slot(s).
}
