% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{member}
\alias{member}
\alias{legco_member}
\title{Members of LegCo}
\usage{
member(member_id = NULL, extra_param = NULL, count = FALSE, verbose = TRUE)

legco_member(
  member_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{member_id}{The id of a LegCo member, or a vector of ids. If `NULL`,
returns results of all LegCo members. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the basic information of LegCo members.
}
\examples{
\dontrun{
# Fetch full list of members
member()
}

}
