% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_term.R
\name{member_term}
\alias{member_term}
\alias{legco_member_term}
\title{Terms of Members of LegCo}
\usage{
member_term(
  member_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)

legco_member_term(
  member_id = NULL,
  extra_param = NULL,
  count = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{member_id}{The id of a LegCo member, or a vector of ids. If `NULL`,
returns results of all LegCo members. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{count}{If `TRUE`, returns only the total count of records that matches
the paramter(s) instead of the result. Defaults to `FALSE`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the serving terms in Term ID of LegCo members
}
\examples{
\dontrun{
#Fetches the term served by Hon Chan Kin-por and Kwong Chun-yu
member_term(id = c(273, 924))
}

}
