% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_term.R
\name{member_term}
\alias{member_term}
\title{Terms of Members of LegCo}
\usage{
member_term(id = NULL, extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{id}{The id of a LegCo member, or a vector of ids. If `NULL`, returns
all LegCo members. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the serving terms in Term ID of LegCo members
}
