% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\alias{legco_membership}
\title{Membership of LegCo Committees}
\usage{
membership(id = NULL, member_id = NULL, committee_id = NULL,
  term_id = NULL, post = NULL, n = 10000, extra_param = NULL,
  verbose = TRUE)

legco_membership(id = NULL, member_id = NULL, committee_id = NULL,
  term_id = NULL, post = NULL, n = 10000, extra_param = NULL,
  verbose = TRUE)
}
\arguments{
\item{id}{The id of a committee membership, or a vector of ids. If `NULL`, returns
all committee membership. Defaults to `NULL`.}

\item{member_id}{The code of a LegCo member, or a vector of ids. If `NULL`,
returns all LegCo members. Defaults to `NULL`.}

\item{committee_id}{The code of a LegCo committee, or a vector of ids. If
`NULL`, returns records of all LegCo committees. Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
membership of committees of all terms.}

\item{post}{The post of a member in the committee, or a vector of posts.
`'President'`, `'Chairman'`, `'Deputy Chairman'`, `'Member'` and
`'Convenor'` return members of the respective post. If `NULL`, returns
members of all posts. Defaults to `NULL`.}

\item{n}{The number of records to request. Defaults to `10000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the members of LegCo committees.
}
