% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session}
\alias{session}
\alias{legco_session}
\title{Session of LegCo}
\usage{
session(id = NULL, term_id = NULL, date = NULL, extra_param = NULL,
  verbose = TRUE)

legco_session(id = NULL, term_id = NULL, date = NULL,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{id}{The id of a session, or a vector of ids. If `NULL`, returns terms.
Defaults to `NULL`.}

\item{term_id}{The id of a term, or a vector of ids. If `NULL`, returns
sessions of all terms.}

\item{date}{Only fetch the session in which the specified date falls within.
Accepts character values in `'YYYY-MM-DD'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a date with `as.Date()`. Defaults to `NULL`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch the basic information of LegCo sessions.
}
