% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speakers.R
\name{speakers}
\alias{speakers}
\alias{legco_speakers}
\title{Speakers at LegCo}
\usage{
speakers(id = NULL, type = "all", extra_args = NULL,
  verbose = TRUE)

legco_speakers(id = NULL, type = "all", extra_args = NULL,
  verbose = TRUE)
}
\arguments{
\item{id}{The id of a speaker at the Legislative Council. If `NULL`,
returns a list of all speakers. Default to `NULL`.}

\item{type}{The position of a speaker. `'all'` returns all speakers.
`'PO'` returns public officers. `'LC'` returns key appointment holders and staff
at LegCo, such as President, Chariman and clerk.`'MB'` returns LegCo members.
Default to `'all'`.}

\item{extra_args}{Additional query string options defined in LegCo API.
Start with `'&'` then followed by the option.}

\item{verbose}{Default to `TRUE`.}
}
\description{
Fetch basic details of the speakers at LegCo, including LegCo members,
government officials and Secretariat staff.
}
