% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_record.R
\name{voting_record}
\alias{voting_record}
\alias{legco_voting_record}
\title{Voting Record in LegCo Meetings}
\usage{
voting_record(committee = NULL, term_id = NULL, result = "all",
  name_ch = NULL, seperate_mechanism = NULL,
  from = "1900-01-01T00:00:00", to = Sys.time(), n = 10000,
  extra_param = NULL, verbose = TRUE)

legco_voting_record(committee = NULL, term_id = NULL, result = "all",
  name_ch = NULL, seperate_mechanism = NULL,
  from = "1900-01-01T00:00:00", to = Sys.time(), n = 10000,
  extra_param = NULL, verbose = TRUE)
}
\arguments{
\item{committee}{The name of the committee or subcommittee. Defaults to
`NULL`.}

\item{term_id}{The ID of a term. Defaults to `NULL`.}

\item{result}{The voting result. `'passed'` returns motions that have been
passed. `'vetoed'` returns motions that have been vetoed. `'all'` returns
all motions that has been voted in LegCo. Defaults to `all`.}

\item{name_ch}{The name of a LegCo member, or a vector of names. If `NULL`,
returns voting records  of all members. Defaults to `NULL`.}

\item{seperate_mechanism}{Only fetch votes that were counted with the vote
seperate mechanism, i.e. requiring majority in both geographical and
functional constituencies to pass. If `NULL`, returns all votes regardless
of the vote counting mechanism used. Defaults to `NULL`.}

\item{from}{Only fetch votes conducted at or after this time. Accepts
character values in `'YYYY-MM-DDTHH:MM:SS'` format, and objects of class
`Date`, `POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced
to a time with `as.POSIXlt()`. Defaults to `'1900-01-01T00:00:00'`.}

\item{to}{Only fetch votes conducted at or before this time. Accepts character
values in `'YYYY-MM-DDTHH:MM:SS'` format, and objects of class `Date`,
`POSIXt`, `POSIXct`, `POSIXlt` or anything else that can be coerced to a
time with `as.POSIXlt()`. Defaults to system time.}

\item{n}{The number of records to request. Defaults to `10000`.}

\item{extra_param}{Additional query parameters defined in LegCo API. Must
begin with `'&'`.}

\item{verbose}{Defaults to `TRUE`.}
}
\description{
Fetch voting records of LegCo council, the Finance Committee and its
subcommittees and the House Committee meetings.
}
